+++ docs/source/speculative-execution/simple.md
@@ -17,7 +17,7 @@ use scylla::{
     Session,
     SessionBuilder,
     speculative_execution::SimpleSpeculativeExecutionPolicy,
-    transport::execution_profile::ExecutionProfile,
+    execution::ExecutionProfile,
 };
 
 let policy = SimpleSpeculativeExecutionPolicy {
