+++ examples/speculative-execution.rs
@@ -1,6 +1,6 @@
 use scylla::{
-    speculative_execution::PercentileSpeculativeExecutionPolicy,
-    transport::execution_profile::ExecutionProfile, Session, SessionBuilder,
+    execution::ExecutionProfile, speculative_execution::PercentileSpeculativeExecutionPolicy,
+    Session, SessionBuilder,
 };
 
 use anyhow::Result;
