+++ scylla/src/execution/execution_profile.rs
@@ -18,7 +18,7 @@
 //! # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
 //! use scylla::{Session, SessionBuilder};
 //! use scylla::statement::Consistency;
-//! use scylla::transport::ExecutionProfile;
+//! use scylla::execution::ExecutionProfile;
 //!
 //! let profile = ExecutionProfile::builder()
 //!     .consistency(Consistency::LocalOne)
@@ -44,7 +44,7 @@
 //! # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
 //! use scylla::query::Query;
 //! use scylla::statement::Consistency;
-//! use scylla::transport::ExecutionProfile;
+//! use scylla::execution::ExecutionProfile;
 //! use std::time::Duration;
 //!
 //! let profile = ExecutionProfile::builder()
@@ -71,7 +71,7 @@
 //! # use std::error::Error;
 //! # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
 //! use scylla::statement::Consistency;
-//! use scylla::transport::ExecutionProfile;
+//! use scylla::execution::ExecutionProfile;
 //! use std::time::Duration;
 //!
 //! let base_profile = ExecutionProfile::builder()
@@ -112,7 +112,7 @@
 //! use scylla::{Session, SessionBuilder};
 //! use scylla::query::Query;
 //! use scylla::statement::Consistency;
-//! use scylla::transport::ExecutionProfile;
+//! use scylla::execution::ExecutionProfile;
 //!
 //! let profile1 = ExecutionProfile::builder()
 //!     .consistency(Consistency::One)
@@ -171,10 +171,10 @@ use crate::{
 };
 
 pub(crate) mod defaults {
+    use crate::execution::ExecutionProfileInner;
     use crate::load_balancing::{self, LoadBalancingPolicy};
     use crate::retry_policy::{DefaultRetryPolicy, RetryPolicy};
     use crate::speculative_execution::SpeculativeExecutionPolicy;
-    use crate::transport::execution_profile::ExecutionProfileInner;
     use scylla_cql::frame::types::SerialConsistency;
     use scylla_cql::Consistency;
     use std::sync::Arc;
@@ -216,7 +216,7 @@ pub(crate) mod defaults {
 /// # Example
 ///
 /// ```
-/// # use scylla::transport::{ExecutionProfile, retry_policy::FallthroughRetryPolicy};
+/// # use scylla::execution::{ExecutionProfile, retries::FallthroughRetryPolicy};
 /// # use scylla::statement::Consistency;
 /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
 /// let profile: ExecutionProfile = ExecutionProfile::builder()
@@ -242,7 +242,7 @@ impl ExecutionProfileBuilder {
     ///
     /// # Example
     /// ```
-    /// # use scylla::transport::ExecutionProfile;
+    /// # use scylla::execution::ExecutionProfile;
     /// # use std::time::Duration;
     /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
     /// let profile: ExecutionProfile = ExecutionProfile::builder()
@@ -276,8 +276,8 @@ impl ExecutionProfileBuilder {
     ///
     /// # Example
     /// ```
-    /// # use scylla::transport::ExecutionProfile;
-    /// # use scylla::transport::load_balancing::DefaultPolicy;
+    /// # use scylla::execution::ExecutionProfile;
+    /// # use scylla::execution::load_balancing::DefaultPolicy;
     /// # use std::sync::Arc;
     /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
     /// let profile: ExecutionProfile = ExecutionProfile::builder()
@@ -300,8 +300,8 @@ impl ExecutionProfileBuilder {
     ///
     /// # Example
     /// ```
-    /// use scylla::transport::retry_policy::DefaultRetryPolicy;
-    /// # use scylla::transport::ExecutionProfile;
+    /// use scylla::execution::retries::DefaultRetryPolicy;
+    /// # use scylla::execution::ExecutionProfile;
     /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
     /// let profile: ExecutionProfile = ExecutionProfile::builder()
     ///     .retry_policy(Box::new(DefaultRetryPolicy::new()))
@@ -323,8 +323,8 @@ impl ExecutionProfileBuilder {
     /// # fn check_only_compiles() -> Result<(), Box<dyn Error>> {
     /// use std::{sync::Arc, time::Duration};
     /// use scylla::{
-    ///     transport::ExecutionProfile,
-    ///     transport::speculative_execution::SimpleSpeculativeExecutionPolicy,
+    ///     execution::ExecutionProfile,
+    ///     execution::speculative_execution::SimpleSpeculativeExecutionPolicy,
     /// };
     ///
     /// let policy = SimpleSpeculativeExecutionPolicy {
@@ -350,8 +350,8 @@ impl ExecutionProfileBuilder {
     ///
     /// # Example
     /// ```
-    /// use scylla::transport::retry_policy::DefaultRetryPolicy;
-    /// # use scylla::transport::ExecutionProfile;
+    /// use scylla::execution::retries::DefaultRetryPolicy;
+    /// # use scylla::execution::ExecutionProfile;
     /// # fn example() -> Result<(), Box<dyn std::error::Error>> {
     /// let profile: ExecutionProfile = ExecutionProfile::builder()
     ///     .retry_policy(Box::new(DefaultRetryPolicy::new()))
