+++ scylla/src/execution/mod.rs
@@ -1,1 +1,8 @@
+mod execution_profile;
 pub mod load_balancing;
+
+pub(crate) use execution_profile::ExecutionProfileInner;
+pub use execution_profile::{ExecutionProfile, ExecutionProfileBuilder, ExecutionProfileHandle};
+
+#[cfg(test)]
+pub(crate) use execution_profile::defaults;
