+++ scylla/src/statement/prepared_statement.rs
@@ -10,13 +10,13 @@ use uuid::Uuid;
 use scylla_cql::frame::response::result::ColumnSpec;
 
 use super::StatementConfig;
+use crate::execution::ExecutionProfileHandle;
 use crate::frame::response::result::PreparedMetadata;
 use crate::frame::types::{Consistency, SerialConsistency};
 use crate::frame::value::SerializedValues;
 use crate::history::HistoryListener;
 use crate::retry_policy::RetryPolicy;
 use crate::sharding::Token;
-use crate::transport::execution_profile::ExecutionProfileHandle;
 use crate::transport::partitioner::{Partitioner, PartitionerHasher, PartitionerName};
 
 /// Represents a statement prepared on the server.
