+++ scylla/src/transport/mod.rs
@@ -1,5 +1,4 @@
 pub mod downgrading_consistency_retry_policy;
-pub mod execution_profile;
 pub mod iterator;
 pub(crate) mod metrics;
 pub mod partitioner;
@@ -8,7 +7,6 @@ pub mod retry_policy;
 pub mod speculative_execution;
 
 pub use crate::frame::{Authenticator, Compression};
-pub use execution_profile::ExecutionProfile;
 pub use scylla_cql::errors;
 
 #[cfg(test)]
